==> src/competencia_02_stacking_exp_colaborativo/5. jIL_dataframe_train_final.ipynb <==
## A. Configuraciones Generales.

#1. Librerías.
%run "./librerias.ipynb"

#2. Constantes.
#i. General.
%run "./constantes.ipynb"

#ii. Particular.
#a. Meses de Train.
mes_train =  mes_train_ult_anio
#b. Cantidad de meses de Train.
cantidad_meses_train = 12
#c. Output del dataset mergeado de los modelos individuales.
dataset_joineado_stacking = datasets_path + "competencia_02_stacking.parquet"

#3. Funciones
%run "./funciones.ipynb"

#4. Lectura de datos.
#i. Modelo 1.
m1 = pd.read_parquet(path_output_m1)
#ii. Modelo 2.
m2 = pd.read_parquet(path_output_m2)
#iii. Modelo 3.
m3 = pd.read_parquet(path_output_m3)
#iv. Modelo 4.
m4 = pd.read_parquet(path_output_m4)
#v. Leo todos los datos para traerme el target real de dichos datos.
target = pd.read_parquet(dataset_file_fe_12_3)

## B. Mergeo y Exportación del dataframe de Train final.

#1. Filtro los meses de interés.
target = target[target["foto_mes"].isin(mes_train_ult_anio)]

#2. Me quedo solamente con el target.
target = target[["numero_de_cliente","foto_mes","clase_ternaria"]]

#3. Mergeo toda la información en un solo dataframe.
#i. Defino las columnas de mergeo.
cols_mergeo = ["numero_de_cliente","foto_mes"]
#ii. Mergeo propiamente dicho.
data = m1.merge(m2,how="left",on=cols_mergeo).merge(m3,how="left",on=cols_mergeo).merge(m4,how="left",on=cols_mergeo).merge(target,how="left",on=cols_mergeo)

#4. Exporto el dataframe final.
data.to_parquet(dataset_joineado_stacking)
